ai_alert_system/requirements.txt
# AI智能告警系统核心依赖包
langchain==0.1.0
langgraph==0.0.26
langchain-openai==0.0.8
langchain-community==0.0.10
chromadb==0.4.18
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
grpcio==1.60.0
grpcio-tools==1.60.0
structlog==23.2.0
rich==13.7.0
asyncio-mqtt==0.16.1
